+++ src/main/javascript/Core/AppEventHandlers.js
@@ -7,8 +7,9 @@ import { handleOutgoingEvent } from './EventHandler';
 
 /**
  * @method
+ * @param {App} app
  */
-export const registerEventHandlers = () => {
-  handleOutgoingEvent(Events.EVENT_RESET_SIZE, Events.props.EVENT_RESET_SIZE);
-  handleOutgoingEvent(Events.EVENT_SUBSCRIBE, Events.props.EVENT_SUBSCRIBE);
+export const registerEventHandlers = (app) => {
+  handleOutgoingEvent(app, Events.EVENT_RESET_SIZE, Events.props.EVENT_RESET_SIZE);
+  handleOutgoingEvent(app, Events.EVENT_SUBSCRIBE, Events.props.EVENT_SUBSCRIBE);
 };