+++ src/main/javascript/Core/ContextEventHandlers.js
@@ -7,11 +7,11 @@ import { handleOutgoingEvent } from './EventHandler';
 
 /**
  * @method
- *
+ * @param {App} app
  */
-export const registerEventHandlers = () => {
-  handleOutgoingEvent(Events.EVENT_ME_GET, Events.props.EVENT_ME_GET);
-  handleOutgoingEvent(Events.EVENT_TAB_DATA, Events.props.EVENT_TAB_DATA);
-  handleOutgoingEvent(Events.EVENT_TAB_ACTIVATE, Events.props.EVENT_TAB_ACTIVATE);
-  handleOutgoingEvent(Events.EVENT_TAB_CLOSE, Events.props.EVENT_TAB_CLOSE);
+export const registerEventHandlers = (app) => {
+  handleOutgoingEvent(app, Events.EVENT_ME_GET, Events.props.EVENT_ME_GET);
+  handleOutgoingEvent(app, Events.EVENT_TAB_DATA, Events.props.EVENT_TAB_DATA);
+  handleOutgoingEvent(app, Events.EVENT_TAB_ACTIVATE, Events.props.EVENT_TAB_ACTIVATE);
+  handleOutgoingEvent(app, Events.EVENT_TAB_CLOSE, Events.props.EVENT_TAB_CLOSE);
 };