+++ src/main/javascript/DeskproWindow/EventHandlers.js
@@ -7,8 +7,9 @@ import { handleOutgoingEvent } from '../Core/EventHandler';
 
 /**
  * @method
+ * @param {App} app
  */
-export const registerEventHandlers = () => {
-  handleOutgoingEvent(Events.EVENT_DESKPROWINDOW_INSERT_MARKUP, Events.props.EVENT_DESKPROWINDOW_INSERT_MARKUP);
-  handleOutgoingEvent(Events.EVENT_DESKPROWINDOW_SHOW_NOTIFICATION, Events.props.EVENT_DESKPROWINDOW_SHOW_NOTIFICATION);
+export const registerEventHandlers = (app) => {
+  handleOutgoingEvent(app, Events.EVENT_DESKPROWINDOW_INSERT_MARKUP, Events.props.EVENT_DESKPROWINDOW_INSERT_MARKUP);
+  handleOutgoingEvent(app, Events.EVENT_DESKPROWINDOW_SHOW_NOTIFICATION, Events.props.EVENT_DESKPROWINDOW_SHOW_NOTIFICATION);
 };