+++ src/main/javascript/Security/EventHandlers.js
@@ -8,8 +8,9 @@ import { handleOutgoingEvent } from '../Core/EventHandler';
 
 /**
  * @method
+ * @param {App} app
  */
-export const registerEventHandlers = () => {
-  handleOutgoingEvent(Events.EVENT_SECURITY_AUTHENTICATE_OAUTH, Events.props.EVENT_SECURITY_AUTHENTICATE_OAUTH);
-  handleOutgoingEvent(Events.EVENT_SECURITY_SETTINGS_OAUTH, Events.props.EVENT_SECURITY_SETTINGS_OAUTH);
+export const registerEventHandlers = (app) => {
+  handleOutgoingEvent(app, Events.EVENT_SECURITY_AUTHENTICATE_OAUTH, Events.props.EVENT_SECURITY_AUTHENTICATE_OAUTH);
+  handleOutgoingEvent(app, Events.EVENT_SECURITY_SETTINGS_OAUTH, Events.props.EVENT_SECURITY_SETTINGS_OAUTH);
 };