+++ src/main/javascript/WebAPI/EventHandlers.js
@@ -8,8 +8,9 @@ import { handleOutgoingEvent } from '../Core/EventHandler';
 
 /**
  * @method
+ * @param {App} app
  */
-export const registerEventHandlers = () => {
-  handleOutgoingEvent(Events.EVENT_WEBAPI_REQUEST_DESKPRO, Events.props.EVENT_WEBAPI_REQUEST_DESKPRO);
-  handleOutgoingEvent(Events.EVENT_WEBAPI_REQUEST_FETCH, Events.props.EVENT_WEBAPI_REQUEST_FETCH);
+export const registerEventHandlers = (app) => {
+  handleOutgoingEvent(app, Events.EVENT_WEBAPI_REQUEST_DESKPRO, Events.props.EVENT_WEBAPI_REQUEST_DESKPRO);
+  handleOutgoingEvent(app, Events.EVENT_WEBAPI_REQUEST_FETCH, Events.props.EVENT_WEBAPI_REQUEST_FETCH);
 };